refactor: Conflate `WithReversibleState` into `BaseCallable`
`WithReversibleState` was only used by `BaseCallable`. Having it inlined
makes it easier to understand the code.
