+++ app/services/base_services/base_callable.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 #-- copyright
 # OpenProject is an open source project management software.
 # Copyright (C) the OpenProject GmbH
@@ -31,7 +33,7 @@ module BaseServices
     extend ActiveModel::Callbacks
     define_model_callbacks :call
 
-    include ::WithReversibleState
+    around_call :assign_state
 
     def call(*args)
       self.params = extract_options!(args).deep_symbolize_keys
@@ -39,6 +41,17 @@ module BaseServices
       run_callbacks(:call) do
         perform(*args, **params)
       end
+    end
+
+    # Reuse or append state to the service
+    def with_state(state = {})
+      @state = ::Shared::ServiceState.build(state)
+      self
+    end
+
+    # Access to the shared service state.
+    def state
+      @state ||= ::Shared::ServiceState.build
     end
 
     protected
@@ -50,6 +63,14 @@ module BaseServices
     end
 
     private
+
+    # Assign state to the service result obtained after the service call.
+    # Called by an `around_call` callback.
+    def assign_state
+      yield.tap do |service_result|
+        service_result.state = state
+      end
+    end
 
     def extract_options!(args)
       if args.last.is_a?(Hash)
